push

+++ src/main/java/utils/java/FastCode.java
@@ -5,7 +5,7 @@ import java.util.*;
 public class FastCode {
 
     // 10^9 + 7
-    public static final int MOD = 1000000007;
+    public static final long MOD = 1000000007;
 
     // → ↓ ← ↑
     private static final int[][] dir = {{0, 0}, {0, 1}, {0, -1}, {1, 0}, {-1, 0}};
